Refactor Meal Plans and Update Dependencies
-   Updated various dependencies to their latest versions.
-   Enhanced meal plan deletion by adding a confirmation dialog to prevent accidental deletion.
-   Implemented delete functionality for meal plans.
-   Added a delete button to the meal plan item layout.
-   Refactored the favorite logic and UI in `RecipesAdapter` to improve user experience.
-    Added the `deleteMealPlan` method in `SyncingMealPlanDao`.
-   Updated the Google Services configuration.
-   Updated `HomeFragment` actions to use `HomeFragmentDirections`.
-   Updated `MealPlanDao` to add a delete method.
-   Fixed the `ShopingListFragment` to fix a null reference.
-   Add a delete button for the shopping list items with a confirmation.

+++ app/src/main/java/com/maayn/mealmate/presentation/grocery/ShoppingListAdapter.kt
@@ -1,5 +1,6 @@
 package com.maayn.mealmate.presentation.grocery
 
+import android.app.AlertDialog
 import android.graphics.Paint
 import android.view.LayoutInflater
 import android.view.View
@@ -63,11 +64,19 @@ class ShoppingListAdapter(
 
         init {
             btnDeleteIngredient.setOnClickListener {
-                val position = adapterPosition
-                if (position != RecyclerView.NO_POSITION) {
-                    listener.onDeleteClick(position)
-                }
+                AlertDialog.Builder(it.context)
+                    .setTitle("Delete Ingredient")
+                    .setMessage("Are you sure you want to delete this ingredient?")
+                    .setPositiveButton("Delete") { _, _ ->
+                        val position = adapterPosition
+                        if (position != RecyclerView.NO_POSITION) {
+                            listener.onDeleteClick(position)
+                        }
+                    }
+                    .setNegativeButton("Cancel", null)
+                    .show()
             }
+
 
             checkboxIngredient.setOnCheckedChangeListener { _, isChecked ->
                 val position = adapterPosition
